important fix on GZipRotator
With this commit I hope I can fix a problem with logs saving after a while. They used to just disappear and I'm pretty sure it had something to do with logs deletion.
If not, maybe the problem happens because of multiple loggers editing the same file. Either way, I have to investigate more.